# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Daymark memory app API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MemoryCategory = typeof MemoryCategory[keyof typeof MemoryCategory];


export const MemoryCategory = {
  travel: 'travel',
  friends: 'friends',
  family: 'family',
  achievements: 'achievements',
  everyday: 'everyday',
} as const;

export interface Person {
  id: number;
  name: string;
  /** @nullable */
  relationship?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  birthday?: string | null;
  memoriesCount?: number;
}

export interface Memory {
  id: number;
  title: string;
  /** @nullable */
  story?: string | null;
  date: string;
  /** @nullable */
  location?: string | null;
  category: MemoryCategory;
  /** @nullable */
  mood?: string | null;
  giftColor: string;
  ribbon: string;
  /** @nullable */
  sticker?: string | null;
  photoUrls?: string[];
  people?: Person[];
  isKeptClose: boolean;
  createdAt: string;
}

export type MemoryInputCategory = typeof MemoryInputCategory[keyof typeof MemoryInputCategory];


export const MemoryInputCategory = {
  travel: 'travel',
  friends: 'friends',
  family: 'family',
  achievements: 'achievements',
  everyday: 'everyday',
} as const;

export interface MemoryInput {
  /** @minLength 1 */
  title: string;
  story?: string;
  date: string;
  location?: string;
  category: MemoryInputCategory;
  mood?: string;
  giftColor?: string;
  ribbon?: string;
  sticker?: string;
  photoUrls?: string[];
  personIds?: number[];
}

export type MemoryUpdateCategory = typeof MemoryUpdateCategory[keyof typeof MemoryUpdateCategory];


export const MemoryUpdateCategory = {
  travel: 'travel',
  friends: 'friends',
  family: 'family',
  achievements: 'achievements',
  everyday: 'everyday',
} as const;

export interface MemoryUpdate {
  title?: string;
  story?: string;
  date?: string;
  location?: string;
  category?: MemoryUpdateCategory;
  mood?: string;
  giftColor?: string;
  ribbon?: string;
  sticker?: string;
  photoUrls?: string[];
  personIds?: number[];
}

export interface PersonDetail {
  id: number;
  name: string;
  /** @nullable */
  relationship?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  birthday?: string | null;
  memoriesCount?: number;
  /** @nullable */
  nextImportantDate?: string | null;
  memories: Memory[];
}

export interface PersonInput {
  /** @minLength 1 */
  name: string;
  relationship?: string;
  avatarUrl?: string;
  birthday?: string;
}

export type CalendarEventType = typeof CalendarEventType[keyof typeof CalendarEventType];


export const CalendarEventType = {
  birthday: 'birthday',
  memory: 'memory',
  travel: 'travel',
  anniversary: 'anniversary',
  achievement: 'achievement',
} as const;

export interface CalendarEvent {
  id: number;
  title: string;
  date: string;
  type: CalendarEventType;
  /** @nullable */
  memoryId?: number | null;
  /** @nullable */
  personId?: number | null;
  /** @nullable */
  daysUntil?: number | null;
}

export interface FutureGift {
  id: number;
  title: string;
  recipientName: string;
  unlockDate: string;
  isLocked: boolean;
  /** @nullable */
  message?: string | null;
  photoUrls?: string[];
  createdAt: string;
}

export interface FutureGiftInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  recipientName: string;
  unlockDate: string;
  message?: string;
  photoUrls?: string[];
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface HomeSummary {
  upcomingEvents: CalendarEvent[];
  giftFromPast: Memory | null;
  recentPeople: Person[];
  totalMemories: number;
}

export interface Notification {
  id: number;
  userId: string;
  type: string;
  title: string;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  relatedMemoryId?: number | null;
  /** @nullable */
  relatedFutureGiftId?: number | null;
  /** @nullable */
  readAt?: string | null;
  createdAt: string;
}

export interface NotificationList {
  notifications: Notification[];
  unreadCount: number;
}

export interface UserProfileUpdate {
  firstName?: string;
  lastName?: string;
  profileImageUrl?: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  onboardingCompleted: boolean;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export type ListMemoriesParams = {
category?: string;
view?: ListMemoriesView;
};

export type ListMemoriesView = typeof ListMemoriesView[keyof typeof ListMemoriesView];


export const ListMemoriesView = {
  grid: 'grid',
  timeline: 'timeline',
  calendar: 'calendar',
  map: 'map',
} as const;

export type ListCalendarEventsParams = {
month?: number;
year?: number;
};

export type MarkNotificationRead200 = {
  success: boolean;
};

export type MarkAllNotificationsRead200 = {
  success: boolean;
};