# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Daymark memory app API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memories:
    get:
      operationId: listMemories
      tags: [memories]
      summary: List all memories
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: view
          in: query
          schema:
            type: string
            enum: [grid, timeline, calendar, map]
      responses:
        "200":
          description: List of memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
    post:
      operationId: createMemory
      tags: [memories]
      summary: Create a new memory (wrap a memory)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryInput"
      responses:
        "201":
          description: Created memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
  /memories/{id}:
    get:
      operationId: getMemory
      tags: [memories]
      summary: Get a single memory by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Memory detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "404":
          description: Not found
    patch:
      operationId: updateMemory
      tags: [memories]
      summary: Update a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryUpdate"
      responses:
        "200":
          description: Updated memory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
    delete:
      operationId: deleteMemory
      tags: [memories]
      summary: Delete a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /memories/{id}/keep-close:
    post:
      operationId: keepMemoryClose
      tags: [memories]
      summary: Mark a memory as kept close (favorited)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
...
  /future-gifts:
    get:
      operationId: listFutureGifts
      tags: [future-gifts]
      summary: List all future gifts
      responses:
        "200":
          description: List of future gifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FutureGift"
    post:
      operationId: createFutureGift
      tags: [future-gifts]
      summary: Create a future gift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FutureGiftInput"
      responses:
        "201":
          description: Created future gift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FutureGift"
  /future-gifts/{id}:
    get:
      operationId: getFutureGift
      tags: [future-gifts]
      summary: Get a future gift (only returns content if unlock date has passed)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Future gift
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FutureGift"
...
  /auth/profile:
    patch:
      operationId: patchUserProfile
      tags: [auth]
      summary: Update authenticated user profile fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
...
components:
  schemas:
...
    Memory:
      type: object
      required: [id, title, date, category, giftColor, ribbon, isKeptClose, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        story:
          type: ["string", "null"]
        date:
          type: string
          format: date
        location:
          type: ["string", "null"]
        category:
          type: string
          enum: [travel, friends, family, achievements, everyday]
        mood:
          type: ["string", "null"]
        giftColor:
          type: string
        ribbon:
          type: string
        sticker:
          type: ["string", "null"]
        photoUrls:
          type: array
          items:
            type: string
        people:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        isKeptClose:
          type: boolean
        createdAt:
          type: string
          format: date-time
    MemoryInput:
      type: object
      required: [title, date, category]
      properties:
        title:
          type: string
          minLength: 1
        story:
          type: string
        date:
          type: string
          format: date
        location:
          type: string
        category:
          type: string
          enum: [travel, friends, family, achievements, everyday]
        mood:
          type: string
        giftColor:
          type: string
        ribbon:
          type: string
        sticker:
          type: string
        photoUrls:
          type: array
          items:
            type: string
        personIds:
          type: array
          items:
            type: number
    MemoryUpdate:
      type: object
      properties:
        title:
          type: string
        story:
          type: string
        date:
          type: string
          format: date
        location:
          type: string
        category:
          type: string
          enum: [travel, friends, family, achievements, everyday]
        mood:
          type: string
        giftColor:
          type: string
        ribbon:
          type: string
        sticker:
          type: string
        photoUrls:
          type: array
          items:
            type: string
        personIds:
          type: array
          items:
            type: number
    Person:
      type: object
      required: [id, name]
      properties:
        id:
          type: number
        name:
          type: string
        relationship:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        birthday:
          type: ["string", "null"]
          format: date
        memoriesCount:
          type: number
...
    FutureGift:
      type: object
      required: [id, title, recipientName, unlockDate, isLocked, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        recipientName:
          type: string
        unlockDate:
          type: string
          format: date
        isLocked:
          type: boolean
        message:
          type: ["string", "null"]
        photoUrls:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    FutureGiftInput:
      type: object
      required: [title, recipientName, unlockDate]
      properties:
        title:
          type: string
          minLength: 1
        recipientName:
          type: string
          minLength: 1
        unlockDate:
          type: string
          format: date
        message:
          type: string
        photoUrls:
          type: array
          items:
            type: string
...
    UserProfileUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        profileImageUrl:
          type: string
    AuthUser:
      type: object
      required: [id, onboardingCompleted]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        onboardingCompleted:
          type: boolean
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'